Split Hitster player tests
